split blockgen files
