+++ packages/core/src/app/block-production/index.ts
@@ -1,1 +1,3 @@
 export * from './merkle-interval-tree'
+export * from './state-interval-tree'
+export * from './plasma-block-tree'
